% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageScrambleR-utils.R
\name{imageScrambleR-conversions}
\alias{imageScrambleR-conversions}
\alias{arr2magick}
\alias{magick2arr}
\title{Conversion Between 2-D Array and Image

These two convenience functions provide conversion between images as handled
by \link[package]{magick} and two-dimensional arrays containing grayscale
data for images.}
\usage{
arr2magick(imgArr)

magick2arr(img)
}
\arguments{
\item{imgArr}{a 2-D array containing grayscale pixel values}

\item{img}{a \pkg{magick} image object as returned by
\code{\link[magick:image_read]{magick::image_read()}}}
}
\value{
A 2-D array containing grayscale pixel values, or a \pkg{magick}
    image object containing a grayscale image
}
\description{
Conversion Between 2-D Array and Image

These two convenience functions provide conversion between images as handled
by \link[package]{magick} and two-dimensional arrays containing grayscale
data for images.
}
\examples{
if(interactive()){
    imgArr <- array(as.raw(c(1,1,2,2,1,1,2,2,3,3,4,4,3,3,4,4)),
                    dim = c(4,4))
    print(imgArr)
    magick_imgArr <- arr2magick(imgArr)
    print(magick_imgArr)

    magick_R <- magick::image_read(file.path(R.home("doc"), "html",
                                             "logo.jpg"))
    print(magick_R)
    imgArr_R <- magick2arr(magick_R)
    print(imgArr_R[1:10, 1:10])
}
}
