% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageScrambleR-package.R
\docType{package}
\name{imageScrambleR-package}
\alias{imageScrambleR-package}
\title{Image Scrambling Exercise in R and C++}
\description{
Transposes chunks of an image array to create a scrambled image.
    Scrambling functions are written in pure R and Rcpp-enabled C++ to allow
    benchmark comparisons.
}
\author{
Carrie Minette

Maintainer: Carrie Minette <carrie.minette@coyotes.usd.edu>
}
