% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrambler_naiveR.R
\name{scrambler_naiveR}
\alias{scrambler_naiveR}
\title{Scramble a 2-D Array}
\usage{
scrambler_naiveR(arr, nchunks)
}
\arguments{
\item{arr}{The raw array to be scrambled}

\item{nchunks}{The number of chunks into which each dimension should be
divided, i.e. \code{nchunks = 3} will divide the rows into 3 chunks and
the columns into 3 chunks, yielding 9 pieces}
}
\value{
A scrambled version of \code{arr}
}
\description{
Divides a two-dimensional array into \code{nchunk * nchunk} pieces and
scrambles the order in which the pieces appear without altering the size of
the array.  This implementation is designed to mimic code written by someone
who is not very familiar with vector optimization in R.
}
\examples{
# for reproducible results, call set.seed()
set.seed(42)
scrambleMe <- array(as.raw(c(1,1,2,2,1,1,2,2,3,3,4,4,3,3,4,4)),
                    dim = c(4,4))
print(scrambleMe)
scrambleMe <- scrambler_naiveR(arr = scrambleMe, nchunks = 2)
print(scrambleMe)
}
